Enhancing Code Maintainability: Refactor of `executive.rs`
