+++ core/src/observer/gasman.rs
@@ -1,6 +1,6 @@
 use super::VmObserve;
 use crate::{
-    executive::ExecutiveResult,
+    executive::FrameReturn,
     vm::{ActionParams, Result as VmResult},
 };
 use cfx_parameters::{
@@ -105,7 +105,7 @@ impl VmObserve for GasMan {
         self.record_call_create(&params.gas, cross_space_internal);
     }
 
-    fn record_call_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_call_result(&mut self, result: &VmResult<FrameReturn>) {
         let gas_left =
             result.as_ref().map_or(U256::zero(), |r| r.gas_left.clone());
         self.record_return(&gas_left);
@@ -115,7 +115,7 @@ impl VmObserve for GasMan {
         self.record_call_create(&params.gas, false);
     }
 
-    fn record_create_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_create_result(&mut self, result: &VmResult<FrameReturn>) {
         let gas_left =
             result.as_ref().map_or(U256::zero(), |r| r.gas_left.clone());
         self.record_return(&gas_left);
