+++ core/src/observer/mod.rs
@@ -3,7 +3,7 @@
 // See http://www.gnu.org/licenses/
 
 use crate::{
-    executive::ExecutiveResult,
+    executive::FrameReturn,
     vm::{ActionParams, Result as VmResult},
 };
 pub use cfx_state::tracer::{AddressPocket, StateTracer};
@@ -25,24 +25,24 @@ pub trait VmObserve: StateTracer {
     fn record_call(&mut self, params: &ActionParams);
 
     /// Prepares call result trace
-    fn record_call_result(&mut self, result: &VmResult<ExecutiveResult>);
+    fn record_call_result(&mut self, result: &VmResult<FrameReturn>);
 
     /// Prepares create trace for given params.
     fn record_create(&mut self, params: &ActionParams);
 
     /// Prepares create result trace
-    fn record_create_result(&mut self, result: &VmResult<ExecutiveResult>);
+    fn record_create_result(&mut self, result: &VmResult<FrameReturn>);
 }
 
 /// Nonoperative observer. Does not trace anything.
 impl VmObserve for () {
     fn record_call(&mut self, _: &ActionParams) {}
 
-    fn record_call_result(&mut self, _: &VmResult<ExecutiveResult>) {}
+    fn record_call_result(&mut self, _: &VmResult<FrameReturn>) {}
 
     fn record_create(&mut self, _: &ActionParams) {}
 
-    fn record_create_result(&mut self, _: &VmResult<ExecutiveResult>) {}
+    fn record_create_result(&mut self, _: &VmResult<FrameReturn>) {}
 }
 
 impl<T> VmObserve for &mut T
@@ -52,7 +52,7 @@ where T: VmObserve
         (*self).record_call(params);
     }
 
-    fn record_call_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_call_result(&mut self, result: &VmResult<FrameReturn>) {
         (*self).record_call_result(result);
     }
 
@@ -60,7 +60,7 @@ where T: VmObserve
         (*self).record_create(params);
     }
 
-    fn record_create_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_create_result(&mut self, result: &VmResult<FrameReturn>) {
         (*self).record_create_result(result);
     }
 }
@@ -75,7 +75,7 @@ where
         self.1.record_call(params);
     }
 
-    fn record_call_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_call_result(&mut self, result: &VmResult<FrameReturn>) {
         self.0.record_call_result(result);
         self.1.record_call_result(result);
     }
@@ -85,7 +85,7 @@ where
         self.1.record_create(params);
     }
 
-    fn record_create_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_create_result(&mut self, result: &VmResult<FrameReturn>) {
         self.0.record_create_result(result);
         self.1.record_create_result(result);
     }
