+++ core/src/observer/tracer.rs
@@ -1,6 +1,6 @@
 use super::VmObserve;
 use crate::{
-    executive::ExecutiveResult,
+    executive::FrameReturn,
     observer::trace::{
         Action, Call, CallResult, Create, CreateResult, ExecTrace,
         InternalTransferAction,
@@ -52,11 +52,11 @@ impl VmObserve for ExecutiveTracer {
         self.traces.push(action);
     }
 
-    fn record_call_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_call_result(&mut self, result: &VmResult<FrameReturn>) {
         let action = Action::CallResult(CallResult::from(result));
         let success = matches!(
             result,
-            Ok(ExecutiveResult {
+            Ok(FrameReturn {
                 apply_state: true, ..
             })
         );
@@ -78,11 +78,11 @@ impl VmObserve for ExecutiveTracer {
         self.traces.push(action);
     }
 
-    fn record_create_result(&mut self, result: &VmResult<ExecutiveResult>) {
+    fn record_create_result(&mut self, result: &VmResult<FrameReturn>) {
         let action = Action::CreateResult(CreateResult::from(result));
         let success = matches!(
             result,
-            Ok(ExecutiveResult {
+            Ok(FrameReturn {
                 apply_state: true, ..
             })
         );
