+++ core/src/state/trace.rs
@@ -1,8 +1,10 @@
-use cfx_state::tracer::{AddressPocket::*, StateTracer};
+use cfx_state::tracer::AddressPocket::*;
 use cfx_types::{address_util::AddressUtil, Address, AddressSpaceUtil, U256};
 
+use crate::observer::VmObserve;
+
 pub fn trace_convert_stroage_points(
-    tracer: &mut dyn StateTracer, addr: Address, from_balance: U256,
+    tracer: &mut dyn VmObserve, addr: Address, from_balance: U256,
     from_collateral: U256,
 )
 {
@@ -23,7 +25,7 @@ pub fn trace_convert_stroage_points(
 }
 
 pub fn trace_refund_collateral(
-    tracer: &mut dyn StateTracer, addr: Address, by: U256,
+    tracer: &mut dyn VmObserve, addr: Address, by: U256,
 ) {
     if !by.is_zero() {
         tracer.trace_internal_transfer(
@@ -39,7 +41,7 @@ pub fn trace_refund_collateral(
 }
 
 pub fn trace_occupy_collateral(
-    tracer: &mut dyn StateTracer, addr: Address, by: U256,
+    tracer: &mut dyn VmObserve, addr: Address, by: U256,
 ) {
     if !by.is_zero() {
         tracer.trace_internal_transfer(
